Re-organize gin test by modality (#1049)
