+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Improvements
 * Using ruff to enforce existence of public classes' docstrings [PR #1034](https://github.com/catalystneuro/neuroconv/pull/1034)
+* Separated tests that use external data by modality [PR #1049](https://github.com/catalystneuro/neuroconv/pull/1049)
 
 ## v0.6.1 (August 30, 2024)
