+++ tests/test_minimal/test_schemas.py
