+++ tests/test_on_data/ecephys/spikeglx_multi_probe_metadata.json
