+++ tests/test_on_data/ecephys/spikeglx_single_probe_metadata.json
