+++ tests/test_on_data/ecephys/test_aux_interfaces.py
