+++ tests/test_on_data/ecephys/test_plexon_metadata.py
