+++ tests/test_on_data/ecephys/test_spikeglx_metadata.py
