+++ tests/test_on_data/icephys/test_gin_icephys.py
@@ -20,7 +20,7 @@ try:
 except ImportError:
     HAVE_PARAMETERIZED = False
 # Load the configuration for the data tests
-test_config_dict = load_dict_from_file(Path(__file__).parent / "gin_test_config.json")
+test_config_dict = load_dict_from_file(Path(__file__).parent.parent / "gin_test_config.json")
 
 # GIN dataset: https://gin.g-node.org/NeuralEnsemble/ephy_testing_data
 if os.getenv("CI"):
