+++ tests/test_on_data/ophys/fiber_photometry_metadata.yaml
