+++ tests/test_on_data/ophys/test_imaging_interfaces.py
@@ -36,7 +36,7 @@ from neuroconv.tools.testing.data_interface_mixins import (
 )
 
 try:
-    from .setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
+    from ..setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
 except ImportError:
     from setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
