+++ tests/test_on_data/ophys/test_metadata_combinations.py
@@ -6,7 +6,7 @@ from parameterized import param, parameterized
 from neuroconv import NWBConverter
 from neuroconv.datainterfaces import Suite2pSegmentationInterface, TiffImagingInterface
 
-from .setup_paths import OPHYS_DATA_PATH
+from ..setup_paths import OPHYS_DATA_PATH
 
 TiffImagingInterface_source_data = dict(
     file_path=str(OPHYS_DATA_PATH / "imaging_datasets" / "Tif" / "demoMovie.tif"), sampling_frequency=15.0
