+++ tests/test_on_data/ophys/test_segmentation_interfaces.py
@@ -11,9 +11,9 @@ from neuroconv.tools.testing.data_interface_mixins import (
 )
 
 try:
-    from .setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
+    from ..setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
 except ImportError:
-    from setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
+    from ..setup_paths import OPHYS_DATA_PATH, OUTPUT_PATH
 
 
 class TestCaimanSegmentationInterface(SegmentationExtractorInterfaceTestMixin):
