+++ tests/test_on_data/test_yaml/conversion_specifications/GIN_conversion_specification_no_nwbfile_name_or_other_metadata.yml
