+++ tests/test_on_data/test_yaml/test_yaml_conversion_specification_cli.py
@@ -7,8 +7,8 @@ from pynwb import NWBHDF5IO
 
 from neuroconv.tools import deploy_process
 
-from .setup_paths import ECEPHY_DATA_PATH as DATA_PATH
-from .setup_paths import OUTPUT_PATH
+from ..setup_paths import ECEPHY_DATA_PATH as DATA_PATH
+from ..setup_paths import OUTPUT_PATH
 
 
 class TestYAMLConversionSpecification(TestCase):
